Refactor intention module.
